Fix compilation warnings

+++ src/format_cpp_interface.cpp
@@ -456,7 +456,7 @@ public:
   }
 
   void formatFooter() override {
-    for (int i = 0; i < currentNs.size(); ++i) {
+    for (size_t i = 0; i < currentNs.size(); ++i) {
       out() << "}" << std::endl;
     }
     out() << std::endl;
